+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ### 0.1.13
-* use webpack 4
+* add props scrollHeight
 
 ### 0.1.12
 * add react and react-dom as npm peerDependencies